requirements.txt
PyYAML
jsonschema
ebooklib